Rename/Move apollo-example to example/brewery
